+++ test/solo/notifier.test.js
